Port device atomics + reduce/copy __shared__ + reduce isnan for sm_37 (#188)
- atomic_ops.cuh: replace libcu++ cuda::atomic_ref (sm_60 gate) with native
  atomics + CAS loops; 16-bit half/bf16 via 32-bit-word CAS; complex via float[2].
- reduce_utils.cuh atomic_reduce: 32-bit-word CAS for sub-4-byte types (sm_37 has
  no 8/16-bit atomicCAS).
- reduce_ops.cuh: isnan(half/bf16) via float cast.
- copy_general_input/all_reduce: __shared__ arrays of non-trivially-constructible
  types (half/bf16/complex have non-trivial default ctors in 11.4) -> uninitialized
  aligned byte buffers (more __shared__ sites in reduce/scan/sort next).

Part of #187.

+++ x/mlx/mlx/backend/cuda/reduce/reduce_ops.cuh
@@ -77,7 +77,8 @@ struct Min {
         return b;
       }
     } else if constexpr (!cuda::std::is_integral_v<T>) {
-      if (cuda::std::isnan(a) || cuda::std::isnan(b)) {
+      if (cuda::std::isnan(static_cast<float>(a)) ||
+          cuda::std::isnan(static_cast<float>(b))) { // K80 port: half/bf16 isnan
         return cuda::std::numeric_limits<float>::quiet_NaN();
       }
     }
@@ -101,7 +102,8 @@ struct Max {
         return b;
       }
     } else if constexpr (!cuda::std::is_integral_v<T>) {
-      if (cuda::std::isnan(a) || cuda::std::isnan(b)) {
+      if (cuda::std::isnan(static_cast<float>(a)) ||
+          cuda::std::isnan(static_cast<float>(b))) { // K80 port: half/bf16 isnan
         return cuda::std::numeric_limits<float>::quiet_NaN();
       }
     }
